requirements.txt
azure-functions
azure-storage-blob
numpy
opencv-python
pillow